requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
pydantic
gradio
requests
google-generativeai
python-dotenv
pytz
python-telegram-bot